Module discv5::kbucket

Implementation of a Kademlia routing table as used by a single peer participating in a Kademlia DHT.

The entry point for the API of this module is a KBucketsTable.

Pending Insertions

When the bucket associated with the Key of an inserted entry is full but contains disconnected nodes, it accepts a PendingEntry. Pending entries are inserted lazily when their timeout is found to be expired upon querying the KBucketsTable. When that happens, the KBucketsTable records an AppliedPending result which must be consumed by calling take_applied_pending regularly and / or after performing lookup operations like entry and closest_keys.

Re-exports

pub use crate::handler::ConnectionDirection;

pub use super::ConnectionDirection;

Structs

AbsentEntry

An entry that is not present in any bucket.

AppliedPending

The result of applying a pending node to a bucket, possibly replacing an existing node.

ClosestValue

Closest Iterator Output Value

Distance

A distance between two Keys.

EntryRefView

An immutable by-reference view of a bucket entry.

EntryView

A cloned, immutable view of an entry that is either present in a bucket or pending insertion.

IpBucketFilter

IpTableFilter

KBucketsTable

A KBucketsTable represents a Kademlia routing table.

Key

A Key is a cryptographic hash, identifying both the nodes participating in the Kademlia DHT, as well as records stored in the DHT.

Node

A Node in a bucket, representing a peer participating in the Kademlia DHT together with an associated value (e.g. contact information).

NodeRefView

An immutable by-reference view of a Node.

NodeStatus

The status of a node in a bucket.

PendingEntry

An entry waiting for a slot to be available in a bucket.

PredicateKey

A key that can be returned from the closest_keys function, which indicates if the key matches the predicate or not.

PredicateValue

A value being returned from a predicate closest iterator.

PresentEntry

An entry present in a bucket.

Enums

BucketInsertResult

The result of inserting an entry into a bucket.

ConnectionState

The connection state of a node.

Entry

A reference into a single entry of a KBucketsTable.

FailureReason

A reason for failing to update or insert a node into the bucket.

InsertResult

Informs if the record was inserted.

InsertResult

The result of inserting an entry into a bucket.

UpdateResult

The result of performing an update on a kbucket/table.

Constants

MAX_NODES_PER_BUCKET

Maximum number of nodes in a bucket, i.e. the (fixed) k parameter.

Traits

Filter