Struct discv5::metrics::InternalMetrics

pub struct InternalMetrics {
    pub active_sessions: AtomicUsize,
    pub moving_window: u64,
    pub unsolicited_requests_per_window: AtomicUsize,
    pub bytes_sent: AtomicUsize,
    pub bytes_recv: AtomicUsize,
}

A collection of metrics used throughout the server.

Fields
active_sessions: AtomicUsize

The number of active UDP sessions that are currently established.

moving_window: u64

The number of seconds to store received packets to taking a moving average over.

unsolicited_requests_per_window: AtomicUsize

The number of unsolicited requests received per moving window.

bytes_sent: AtomicUsize

The number of bytes sent.

bytes_recv: AtomicUsize

The number of bytes received.

Implementations

impl InternalMetrics

pub fn add_recv_bytes(&self, bytes: usize)

pub fn add_sent_bytes(&self, bytes: usize)