Module discv5::service

The Discovery v5 protocol. See lib.rs for further details.

Note: Discovered ENR’s are not automatically added to the routing table. Only established sessions get added, ensuring only valid ENRs are added. Manual additions can be made using the add_enr() function.

Response to queries return PeerId. Only the trusted (a session has been established with) PeerId’s are returned, as ENR’s for these PeerId’s are stored in the routing table and as such should have an address to connect to. Untrusted PeerId’s can be obtained from the Service::Discovered event, which is fired as peers get discovered.

Note that although the ENR crate does support Ed25519 keys, these are currently not supported as the ECDH procedure isn’t specified in the specification. Therefore, only secp256k1 keys are supported currently.

Structs

Pong

Service

TalkRequest

Request type for Protocols using TalkReq message.

Enums

CallbackResponse

The kinds of responses we can send back to the discv5 layer.

QueryKind

The types of queries that can be made.

ServiceRequest

The types of requests to send to the Discv5 service.