Enum discv5::socket::ListenConfig

pub enum ListenConfig {
    Ipv4 {
        ip: Ipv4Addr,
        port: u16,
    },
    Ipv6 {
        ip: Ipv6Addr,
        port: u16,
    },
    DualStack {
        ipv4: Ipv4Addr,
        ipv4_port: u16,
        ipv6: Ipv6Addr,
        ipv6_port: u16,
    },
}

Configuration for the sockets to listen on.

Default implementation is the UNSPECIFIED ipv4 address with port 9000.

Variants

Ipv4

Fields

ip: Ipv4Addr

port: u16

Ipv6

Fields

ip: Ipv6Addr

port: u16

DualStack

Fields

ipv4: Ipv4Addr

ipv4_port: u16

ipv6: Ipv6Addr

ipv6_port: u16

Implementations

impl ListenConfig

pub fn from_ip(ip: IpAddr, port: u16) -> ListenConfig

If an IpAddr is known, a ListenConfig can be created based on the version. This will not create a dual stack configuration.

pub fn from_two_sockets( ipv4: Option<SocketAddrV4>, ipv6: Option<SocketAddrV6> ) -> ListenConfig

Allows optional ipv4 and ipv6 addresses to be entered to create a ListenConfig. If both are specified a dual-stack configuration will result. This will panic if both parameters are None.

pub fn with_ipv4(self, ip: Ipv4Addr, port: u16) -> ListenConfig

Sets an ipv4 socket. This will override any past ipv4 configuration and will promote the configuration to dual socket if an ipv6 socket is configured.

pub fn with_ipv6(self, ip: Ipv6Addr, port: u16) -> ListenConfig

Sets an ipv6 socket. This will override any past ipv6 configuration and will promote the configuration to dual socket if an ipv4 socket is configured.