Struct discv5::ConfigBuilder

pub struct ConfigBuilder { /* private fields */ }

Implementations

impl ConfigBuilder

pub fn new(listen_config: ListenConfig) -> Self

pub fn enable_packet_filter(&mut self) -> &mut Self

Whether to enable the incoming packet filter.

pub fn request_timeout(&mut self, timeout: Duration) -> &mut Self

The request timeout for each UDP request.

pub fn vote_duration(&mut self, vote_duration: Duration) -> &mut Self

The interval over which votes are remembered when determining our external IP. A lower interval will respond faster to IP changes. Default is 30 seconds.

pub fn query_peer_timeout(&mut self, timeout: Duration) -> &mut Self

The timeout after which a QueryPeer in an ongoing query is marked unresponsive. Unresponsive peers don’t count towards the parallelism limits for a query. Hence, we may potentially end up making more requests to good peers.

pub fn query_timeout(&mut self, timeout: Duration) -> &mut Self

The timeout for an entire query. Any peers discovered before this timeout are returned.

pub fn request_retries(&mut self, retries: u8) -> &mut Self

The number of retries for each UDP request.

pub fn session_timeout(&mut self, timeout: Duration) -> &mut Self

The session timeout for each node.

pub fn session_cache_capacity(&mut self, capacity: usize) -> &mut Self

The maximum number of established sessions to maintain.

pub fn disable_enr_update(&mut self) -> &mut Self

Disables the auto-update of the local ENR IP and port based on PONG responses from peers.

pub fn max_nodes_response(&mut self, max: usize) -> &mut Self

The maximum number of nodes we response to a find nodes request.

pub fn enr_peer_update_min(&mut self, min: usize) -> &mut Self

The minimum number of peer’s who agree on an external IP port before updating the local ENR.

pub fn query_parallelism(&mut self, parallelism: usize) -> &mut Self

The number of peers to request in parallel in a single query.

pub fn ip_limit(&mut self) -> &mut Self

Limits the number of IP addresses from the same /24 subnet in the kbuckets table. This is to mitigate eclipse attacks.

pub fn incoming_bucket_limit(&mut self, limit: usize) -> &mut Self

Sets a maximum limit to the number of incoming nodes (nodes that have dialed us) to exist per-bucket. This cannot be larger than the bucket size (16). By default, half of every bucket (8 positions) is the largest number of nodes that we accept that dial us.

pub fn table_filter(&mut self, filter: fn(_: &Enr) -> bool) -> &mut Self

A filter used to decide whether to insert nodes into our local routing table. Nodes can be excluded if they do not pass this filter.

pub fn ping_interval(&mut self, interval: Duration) -> &mut Self

The time between pings to ensure connectivity amongst connected nodes.

pub fn disable_report_discovered_peers(&mut self) -> &mut Self

Disables reporting of discovered peers through the event stream.

pub fn filter_rate_limiter( &mut self, rate_limiter: Option<RateLimiter> ) -> &mut Self

A rate limiter for limiting inbound requests.

pub fn filter_max_nodes_per_ip( &mut self, max_nodes_per_ip: Option<usize> ) -> &mut Self

If the filter is enabled, sets the maximum number of nodes per IP before banning the IP.

pub fn filter_max_bans_per_ip( &mut self, max_bans_per_ip: Option<usize> ) -> &mut Self

The maximum number of times nodes from a single IP can be banned, before the IP itself gets banned.

pub fn permit_ban_list(&mut self, list: PermitBanList) -> &mut Self

A set of lists that permit or ban IP’s or NodeIds from the server. See crate::PermitBanList.

pub fn ban_duration(&mut self, ban_duration: Option<Duration>) -> &mut Self

Set the default duration for which nodes are banned for. This timeouts are checked every 5 minutes, so the precision will be to the nearest 5 minutes. If set to None, bans from the filter will last indefinitely. Default is 1 hour.

pub fn executor( &mut self, executor: Box<dyn Executor + Send + Sync> ) -> &mut Self

A custom executor which can spawn the discv5 tasks. This must be a tokio runtime, with timing support.

pub fn build(&mut self) -> Config